requirements/production.txt
Django==1.11.3
psycopg2
prettyconf
dj-database-url
djangorestframework
django-mptt==0.8.7
drfdocs==0.0.11
django-storages==1.6.5
boto3==1.4.8
djangorestframework-recursive==0.1.2
raven>=3